int_sensor: Add a little hack to solve int_sensor wakelock-fix drain
Adds a very poor hack that converts the sensor_ind wakelock
to a timed one on-the-fly. This is unfortunately necessary
for broken DSP firmwares on old devices, combined with
proprietary sensor binaries which don't account for that.
To mitigate this issue, this configuration option applies
a 5 seconds timeout to this wakelock.

Thx for this commit to jerpelea (org. commit : https://github.com/sonyxperiadev/kernel/pull/1588/files )

Signed-off-by: ThunderStorms21th - nalas